Reorganize components and move authenticate axios call to it own file
